Added deployment script, also changed effects and calculators

+++ src/OpenRpg.Genres.Fantasy/Combat/Damage.cs
@@ -2,10 +2,10 @@ namespace OpenRpg.Genres.Fantasy.Combat
 {
     public class Damage
     {
-        public byte Type { get; }
+        public int Type { get; }
         public float Value { get; }
 
-        public Damage(byte type, float value)
+        public Damage(int type, float value)
         {
             Type = type;
             Value = value;
