+++ src/OpenRpg.Genres.Fantasy/Types/DamageTypes.cs
@@ -2,20 +2,20 @@ namespace OpenRpg.Genres.Fantasy.Types
 {
     public class DamageTypes
     {
-        public static byte UnknownDamage = 0;
+        public static int UnknownDamage = 0;
 
         // Melee types
-        public static byte SlashingDamage = 1;
-        public static byte BluntDamage = 2;
-        public static byte PiercingDamage = 3;
-        public static byte UnarmedDamage = 4;
+        public static int SlashingDamage = 1;
+        public static int BluntDamage = 2;
+        public static int PiercingDamage = 3;
+        public static int UnarmedDamage = 4;
 
         // Magic types
-        public static byte FireDamage = 50;
-        public static byte IceDamage = 51;
-        public static byte EarthDamage = 52;
-        public static byte WindDamage = 53;
-        public static byte LightDamage = 54;
-        public static byte DarkDamage = 55;
+        public static int FireDamage = 50;
+        public static int IceDamage = 51;
+        public static int EarthDamage = 52;
+        public static int WindDamage = 53;
+        public static int LightDamage = 54;
+        public static int DarkDamage = 55;
     }
 }
