+++ src/OpenRpg.Genres.Fantasy/Types/EquipmentTypes.cs
@@ -2,19 +2,19 @@ namespace OpenRpg.Genres.Fantasy.Types
 {
     public class EquipmentSlotTypes
     {
-        public static byte UnknownSlot = 0;
+        public static int UnknownSlot = 0;
         
-        public static byte HeadSlot = 1;
-        public static byte BackSlot = 2;
-        public static byte ShoulderSlot = 3;
-        public static byte UpperArmourSlot = 4;
-        public static byte LowerArmourSlot = 5;
-        public static byte HandsSlot = 6;
-        public static byte FeetSlot = 7;
-        public static byte LeftHandSlot = 8;
-        public static byte RightHandSlot = 9;
-        public static byte NeckSlot = 10;
-        public static byte Finger1Slot = 11;
-        public static byte Finger2Slot = 12;
+        public static int HeadSlot = 1;
+        public static int BackSlot = 2;
+        public static int ShoulderSlot = 3;
+        public static int UpperArmourSlot = 4;
+        public static int LowerArmourSlot = 5;
+        public static int HandsSlot = 6;
+        public static int FeetSlot = 7;
+        public static int LeftHandSlot = 8;
+        public static int RightHandSlot = 9;
+        public static int NeckSlot = 10;
+        public static int Finger1Slot = 11;
+        public static int Finger2Slot = 12;
     }
 }
