+++ src/OpenRpg.Genres.Fantasy/Types/ItemQualityTypes.cs
@@ -2,17 +2,17 @@ namespace OpenRpg.Genres.Fantasy.Types
 {
     public class ItemQualityTypes
     {
-        public static byte UnknownQuality = 0;
+        public static int UnknownQuality = 0;
 
-        public static byte JunkQuality = 1;
-        public static byte CommonQuality = 2;
-        public static byte UncommonQuality = 3;
-        public static byte WellMadeQuality = 4;
-        public static byte RareQuality = 5;
-        public static byte MagicalQuality = 6;
-        public static byte EpicQuality = 7;
-        public static byte IncredibleQuality = 8;
-        public static byte LegendaryQuality = 9;
-        public static byte MythicalQuality = 10;
+        public static int JunkQuality = 1;
+        public static int CommonQuality = 2;
+        public static int UncommonQuality = 3;
+        public static int WellMadeQuality = 4;
+        public static int RareQuality = 5;
+        public static int MagicalQuality = 6;
+        public static int EpicQuality = 7;
+        public static int IncredibleQuality = 8;
+        public static int LegendaryQuality = 9;
+        public static int MythicalQuality = 10;
     }
 }
