+++ src/OpenRpg.Genres.Fantasy/Types/RequirementTypes.cs
@@ -2,24 +2,24 @@ namespace OpenRpg.Genres.Fantasy.Types
 {
     public class RequirementTypes
     {
-        public static byte UnknownRequirement = 0;
+        public static int UnknownRequirement = 0;
         
-        public static byte RaceRequirement = 1;
-        public static byte ClassRequirement = 2;
-        public static byte LevelRequirement = 3;
-        public static byte TriggerRequirement = 4;
-        public static byte QuestCompleteRequirement = 5;
-        public static byte SkillRequirement = 6;
-        public static byte InventoryItemRequirement = 7;
-        public static byte MonsterSightedRequirement = 8;
-        public static byte AreaExploredRequirement = 9;
-        public static byte RegionExploredRequirement = 10;
-        public static byte EquipmentItemRequirement = 11;
-        public static byte StrengthRequirement = 12;
-        public static byte DexterityRequirement = 13;
-        public static byte ConstitutionRequirement = 14;
-        public static byte IntlligenceRequirement = 15;
-        public static byte GenderRequirement = 16;
-        public static byte QuestActiveRequirement = 17;
+        public static int RaceRequirement = 1;
+        public static int ClassRequirement = 2;
+        public static int LevelRequirement = 3;
+        public static int TriggerRequirement = 4;
+        public static int QuestCompleteRequirement = 5;
+        public static int SkillRequirement = 6;
+        public static int InventoryItemRequirement = 7;
+        public static int MonsterSightedRequirement = 8;
+        public static int AreaExploredRequirement = 9;
+        public static int RegionExploredRequirement = 10;
+        public static int EquipmentItemRequirement = 11;
+        public static int StrengthRequirement = 12;
+        public static int DexterityRequirement = 13;
+        public static int ConstitutionRequirement = 14;
+        public static int IntlligenceRequirement = 15;
+        public static int GenderRequirement = 16;
+        public static int QuestActiveRequirement = 17;
     }
 }
