+++ src/OpenRpg.Genres.Fantasy/Types/RewardTypes.cs
@@ -2,12 +2,12 @@ namespace OpenRpg.Genres.Fantasy.Types
 {
     public class RewardTypes
     {
-        public static byte UnknownReward = 0;
+        public static int UnknownReward = 0;
        
-        public static byte ExperienceReward = 1;
-        public static byte GoldReward = 2;
-        public static byte ItemReward = 3;
-        public static byte QuestReward = 4;
-        public static byte TriggerReward = 5;
+        public static int ExperienceReward = 1;
+        public static int GoldReward = 2;
+        public static int ItemReward = 3;
+        public static int QuestReward = 4;
+        public static int TriggerReward = 5;
     }
 }
